requirements.dev.txt
# dev dependencies
ipython>=7.0.0,<=8.0.0
ipdb
pudb
pytest-watch

# Code Quality
black
flake8
pyproject-flake8
isort

# Install project as editable
-e .
